// apps/gateway/src/common/swagger/enums/posts.enum.ts
export enum PostsEnum {
  getById = 'Поиск поста по postId',
  getById_Ok = 'Пост найден по postId',

  deleteById = 'Удаление поста по postId. Пользователь определяется по accessToken',

  updateById = 'Обновление поста по postId. Пользователь определяется по accessToken',

  createPost = 'Создание поста происходит в момент загрузок картинок и дефолтным пустым description. Пользователь определяется по accessToken',
  createPost_Ok = 'Создание поста прошло успешно. Картинки загрузились успешно',

  getUserPosts = 'Получение постов конкретного пользователя по его id. Для пагинации используется курсор. Описание курсора внутри',
  getUserPosts_Ok = 'Получение постов конкретного пользователя по его id прошло успешно. Для следующей страницы нужно передать курсов из view модели и id пользователя',

  getPosts = 'Получение постов. Для пагинации используется курсор. Описание курсора внутри',
  getPosts_Ok = 'Получение постов прошло успешно. Для следующей страницы нужно передать курсов из view модели',
}